+++ CHANGELOG.md
@@ -5,8 +5,8 @@ All notable changes to this project will be documented in this file.
 ## [1.0.0] - 2026-07-24
 
 ### Added
-- Initial release of Linkami: A free reimplementation of linkstore.app.
+- Initial release of Linkoteca: A free reimplementation of linkstore.app.
 - Backend implemented with Python.
 - Frontend built with React.
 - Production-ready deployment setup with `systemd` and `Gunicorn` via an automatic install script.
-- Migration script (`migrate_from_linkstore.py`) to seamlessly transfer saved links from Linkstore to Linkami using API tokens.
+- Migration script (`migrate_from_linkstore.py`) to seamlessly transfer saved links from Linkstore to Linkoteca using API tokens.